[API] Move file to its right folder
